Changed rotation field from int to float
Changed variable type of the rotation field in ImageBox class from int
to float - this prevented the editor from being able to use fractional
values for gib rotation, causing the gib animation to look very choppy

+++ src/com/kartoflane/superluminal/elements/FTLGib.java
@@ -151,7 +151,7 @@ public class FTLGib extends ImageBox implements Serializable, DraggableBox {
 	}
 
 	public void setRotation(float angle) {
-		super.setRotation((int) angle);
+		super.setRotation(angle);
 		animRotation = angle;
 	}
